+++ tests/Unit/Jobs/Jav/XCityIdolFetchItemTest.php
@@ -5,27 +5,15 @@ namespace Tests\Unit\Jobs\Jav;
 use App\Jobs\Jav\XCityIdolFetchItem;
 use App\Models\Idol;
 use App\Models\TemporaryUrl;
-use App\Services\Client\CrawlerClientResponse;
-use App\Services\Client\Domain\ResponseInterface;
 use App\Services\Client\XCrawlerClient;
-use PHPUnit\Framework\MockObject\MockObject;
-use Tests\TestCase;
 
-class XCityIdolFetchItemTest extends TestCase
+class XCityIdolFetchItemTest extends AbstractXCityTest
 {
-    private MockObject|XCrawlerClient $mocker;
-
     private TemporaryUrl $url;
 
     public function setUp(): void
     {
         parent::setUp();
-        app()->bind(ResponseInterface::class, CrawlerClientResponse::class);
-        $this->mocker = $this->getMockBuilder(XCrawlerClient::class)->getMock();
-        $this->mocker->method('init')->willReturnSelf();
-        $this->mocker->method('setHeaders')->willReturnSelf();
-        $this->mocker->method('setContentType')->willReturnSelf();
-        $this->fixtures = __DIR__ . '/../../../Fixtures/XCity';
         $this->url = TemporaryUrl::factory()->create(['url' => $this->faker->url, 'source' => $this->faker->uuid, 'state_code' => $this->faker->uuid]);
     }
